Revert "Move code from command/__init__.py to a properly-named file."
This reverts commit 6b024d7c9735f43eb80389be293183ac77a10e68.
