+++ src/python/twitter/pants/commands/goal.py
@@ -45,7 +45,7 @@ from twitter.pants.base.rcfile import RcFile
 from twitter.pants.base.run_info import RunInfo
 from twitter.pants.base.target import Target, TargetDefinitionException
 from twitter.pants.base.workunit import WorkUnit
-from twitter.pants.commands.command import Command
+from twitter.pants.commands import Command
 from twitter.pants.engine import Engine, GroupEngine
 from twitter.pants.goal import Context, GoalError, Phase
 from twitter.pants.goal import Goal as goal, Group as group
